feat(cache) utilize docker hub as a cache (#43)
* wip

* feat(cache) rename docker images to use kong-build-tools docker hub repo

* feat(cache) make the Docker file more cacheable

* chore(travis) force docker into experimental on travis-ci

* chore(travis) wtf travis-ci?

* chore(cache) delay openssl compilation

* fix(travis) let the docker pull fail silently

* test(rename) rename the test docker container

* chore(travis)

* chore(travis) setup python 3.6 on travis-ci

* feat(cache) cache the test container

* fix(cache) assume we should docker push on failures to introspect

* fix(rpm) don't worry about dpkg on rpm builds

* fix(permissions)

* debug

* permissions

* wtf

* revert

* permissions

* fix(permissions) permission and cache hit fails

* feat(cache)

* feat(cache) dont rebuild the base image

* fix(travis) background the setup task

* fix(travis) permissions

* fix(travis) backgrounding the setup task

* feat(travis) background a few docker pulls

* fix(logic) logic fix

* debug

* fix(rhel) login to the redhat registry

* what the actual fuck travis-ci?

* chore(travis) jobify the travis-ci job

* Trigger notification

* typo

* simplify the yaml

* duplicate the job

* further expand the jobs

* chore(travis) try without a yaml alias

* fix(development) fix the development image

* fix(travis-ci)

* fix(travis) foreground the running minikube setup

* fix(travis) can't use fg

+++ Makefile
@@ -1,3 +1,5 @@
+.PHONY: test build-kong
+
 export SHELL:=/bin/bash
 
 RESTY_IMAGE_BASE?=ubuntu
@@ -22,7 +24,7 @@ KONG_PACKAGE_NAME?="kong"
 KONG_CONFLICTS?="kong-enterprise-edition"
 KONG_LICENSE?="ASL 2.0"
 PRIVATE_REPOSITORY?=true
-KONG_TEST_CONTAINER_NAME?=localhost:5000/kong
+KONG_TEST_CONTAINER_NAME?=localhost:5000/kong-$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG)
 KONG_SOURCE_LOCATION?="$$PWD/../kong/"
 KONG_VERSION?=`echo $(KONG_SOURCE_LOCATION)/kong-*.rockspec | sed 's,.*/,,' | cut -d- -f2`
 RESTY_VERSION ?= `grep RESTY_VERSION $(KONG_SOURCE_LOCATION)/.requirements | awk -F"=" '{print $$2}'`
@@ -45,6 +47,17 @@ RESTY_CONFIG_OPTIONS ?= "--with-cc-opt='-I/tmp/openssl/include' \
 LIBYAML_VERSION ?= 0.2.1
 LYAML_VERSION ?= 6.2.3
 
+update-docker-cache:
+ifneq ($(RESTY_IMAGE_BASE),rhel)
+	-./docker_push_latest_if_changed.py --source kong/kong-build-tools:fpm
+	-./docker_push_latest_if_changed.py --source kong/kong-build-tools:development || true
+	-./docker_push_latest_if_changed.py --source kong/kong-build-tools:test_runner && echo "success!" || docker push kong/kong-build-tools:test_runner
+	-./docker_push_latest_if_changed.py --source kong/kong-build-tools:$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG) && echo "success!" || docker push kong/kong-build-tools:$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG)
+	-./docker_push_latest_if_changed.py --source kong/kong-build-tools:kong-$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG) && echo "success!" || docker push kong/kong-build-tools:kong-$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG)
+	-docker tag $(KONG_TEST_CONTAINER_NAME) kong/kong-build-tools:test-$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG)
+	-./docker_push_latest_if_changed.py --source kong/kong-build-tools:test-$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG) && echo "success!" || docker push kong/kong-build-tools:test-$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG)
+endif
+
 release-kong: test
 	RESTY_IMAGE_BASE=$(RESTY_IMAGE_BASE) \
 	RESTY_IMAGE_TAG=$(RESTY_IMAGE_TAG) \
@@ -55,33 +68,37 @@ release-kong: test
 	PRIVATE_REPOSITORY=$(PRIVATE_REPOSITORY) \
 	./release-kong.sh
 
-clean:
-	docker rmi kong:fpm
-	docker rmi kong:kong-$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG)
-	docker rmi kong:$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG)
-
-development:
+build-development-image:
+ifeq ($(RESTY_IMAGE_TAG),xenial)
+	docker pull kong/kong-build-tools:kong-ubuntu-xenial
 	test -s output/kong-$(KONG_VERSION).xenial.all.deb || make package-kong
 	cp output/kong-$(KONG_VERSION).xenial.all.deb output/kong-$(KONG_VERSION).kong-ubuntu-xenial.all.deb
-	docker inspect --type=image kong:kong-ubuntu-xenial > /dev/null || make build-kong
+	docker inspect --type=image kong/kong-build-tools:kong-ubuntu-xenial > /dev/null || make build-kong
 	docker build \
-	--build-arg RESTY_IMAGE_BASE=kong \
-	--build-arg RESTY_IMAGE_TAG=kong-ubuntu-xenial \
+	--cache-from kong/kong-build-tools:development \
+	--build-arg RESTY_IMAGE_BASE=kong/kong-build-tools \
+	--build-arg RESTY_IMAGE_TAG="kong-ubuntu-xenial" \
 	--build-arg KONG_VERSION=$(KONG_VERSION) \
 	--build-arg KONG_UID=$$(id -u) \
 	--build-arg USER=$$USER \
 	--build-arg RUNAS_USER=$$USER \
 	-f test/Dockerfile.deb \
-	-t kong:development .
+	-t kong/kong-build-tools:development .
+endif
+
+development: build-development-image
+ifeq ($(RESTY_IMAGE_TAG),xenial)
 	- docker-compose stop
 	- docker-compose rm -f
 	USER=$$(id -u) docker-compose up -d && \
 	docker-compose exec kong make dev && \
 	docker-compose exec kong /bin/bash
+endif
 
 package-kong: build-kong
 	docker build -f Dockerfile.fpm \
-	-t kong:fpm .
+	--cache-from kong/kong-build-tools:fpm \
+	-t kong/kong-build-tools:fpm .
 	docker run -t --rm \
 	-v $$PWD/output/build:/tmp/build \
 	-v $$PWD/output:/output \
@@ -91,11 +108,12 @@ package-kong: build-kong
 	-e KONG_LICENSE=$(KONG_LICENSE) \
 	-e RESTY_IMAGE_TAG=$(RESTY_IMAGE_TAG) \
 	-e RESTY_IMAGE_BASE=$(RESTY_IMAGE_BASE) \
-	kong:fpm
+	kong/kong-build-tools:fpm
 
 build-kong:
-	docker inspect --type=image kong:$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG) > /dev/null || make build-base
+	docker inspect --type=image kong/kong-build-tools:$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG) > /dev/null || make build-base
 	docker build -f Dockerfile.kong \
+	--cache-from kong/kong-build-tools:kong-$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG) \
 	--build-arg RESTY_VERSION=$(RESTY_VERSION) \
 	--build-arg RESTY_LUAROCKS_VERSION=$(RESTY_LUAROCKS_VERSION) \
 	--build-arg RESTY_OPENSSL_VERSION=$(RESTY_OPENSSL_VERSION) \
@@ -108,11 +126,11 @@ build-kong:
 	--build-arg EDITION=$(EDITION) \
 	--build-arg KONG_GMP_VERSION=$(KONG_GMP_VERSION) \
 	--build-arg KONG_NETTLE_VERSION=$(KONG_NETTLE_VERSION) \
-	-t kong:kong-$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG) .
+	-t kong/kong-build-tools:kong-$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG) .
 	docker run -it --rm \
 	-v $(KONG_SOURCE_LOCATION):/kong \
 	-v $$PWD/output/build:/output/build \
-	kong:kong-$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG)
+	kong/kong-build-tools:kong-$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG)
 
 build-base:
 ifeq ($(RESTY_IMAGE_BASE),rhel)
@@ -125,15 +143,15 @@ ifeq ($(RESTY_IMAGE_BASE),rhel)
 	--build-arg RESTY_IMAGE_BASE=$(RESTY_IMAGE_BASE) \
 	--build-arg REDHAT_USERNAME=$(REDHAT_USERNAME) \
 	--build-arg REDHAT_PASSWORD=$(REDHAT_PASSWORD) \
-	-t kong:$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG) .
+	-t kong/kong-build-tools:$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG) .
 else
 	docker build -f Dockerfile.$(PACKAGE_TYPE) \
+	--cache-from kong/kong-build-tools:$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG) \
 	--build-arg RESTY_IMAGE_TAG="$(RESTY_IMAGE_TAG)" \
 	--build-arg RESTY_IMAGE_BASE=$(RESTY_IMAGE_BASE) \
-	-t kong:$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG) .
+	-t kong/kong-build-tools:$(RESTY_IMAGE_BASE)-$(RESTY_IMAGE_TAG) .
 endif
 
-.PHONY: test
 test: build_test_container
 	KONG_VERSION=$(KONG_VERSION) \
 	RESTY_IMAGE_BASE=$(RESTY_IMAGE_BASE) \
@@ -142,9 +160,9 @@ test: build_test_container
 	./test/run_tests.sh
 
 run_tests:
-	cd test && docker build -t kong:test_runner -f Dockerfile.test_runner .
+	cd test && docker build -t kong/kong-build-tools:test_runner --cache-from kong/kong-build-tools:test_runner -f Dockerfile.test_runner .
 	docker run -it --network host -e RESTY_VERSION=$(RESTY_VERSION) -e KONG_VERSION=$(KONG_VERSION) -e ADMIN_URI=$(TEST_ADMIN_URI) -e PROXY_URI=$(TEST_PROXY_URI) ubuntu printenv
-	docker run -it --network host -e RESTY_VERSION=$(RESTY_VERSION) -e KONG_VERSION=$(KONG_VERSION) -e ADMIN_URI=$(TEST_ADMIN_URI) -e PROXY_URI=$(TEST_PROXY_URI) kong:test_runner /bin/bash -c "py.test -p no:logging -p no:warnings test_*.tavern.yaml"
+	docker run -it --network host -e RESTY_VERSION=$(RESTY_VERSION) -e KONG_VERSION=$(KONG_VERSION) -e ADMIN_URI=$(TEST_ADMIN_URI) -e PROXY_URI=$(TEST_PROXY_URI) kong/kong-build-tools:test_runner /bin/bash -c "py.test -p no:logging -p no:warnings test_*.tavern.yaml"
 
 develop_tests:
 	docker run -it --network host --rm -e RESTY_VERSION=$(RESTY_VERSION) -e KONG_VERSION=$(KONG_VERSION) \
@@ -152,7 +170,6 @@ develop_tests:
 	-e PROXY_URI="http://`kubectl get nodes --namespace default -o jsonpath='{.items[0].status.addresses[0].address}'`:`kubectl get svc --namespace default kong-kong-proxy -o jsonpath='{.spec.ports[0].nodePort}'`" \
 	-v $$PWD/test:/app \
 	kong:test_runner /bin/bash
-  
 
 build_test_container:
 	RESTY_IMAGE_BASE=$(RESTY_IMAGE_BASE) \
